requirements.txt
Flask
beautifulsoup4
python-dotenv
tqdm
google-api-python-client
google-auth-oauthlib
requests
lxml

